Summer update now available and tested: The focus of this update is better explanation of the existing scores. No new scores have been added, so there is no need to re-upload data. Here’s what’s new:
The plot-heritability function is now switched on by default, and twin- and snp-heritabilities have been mapped for 293 diseases and traits, up from 47. Thank you GWASroc for 92 of these! (issue #13)

The all-SNP scores have been scaled so they also are Z-scores. Before they were given on an arbitrary scale on top of a user-distribution plot. But it's much easier to explain when everything - both top-SNP and all-SNP scores - are presented as normalised Z-scores (issue #15)

A new module prsExplainer has also been introduced www.impute.me/prsExplainer - the purpose of this module is to let people play around with different input-scores in a sandbox setting and see directly how it affects outputs.

A list of media and online resources have been added to the ‘about’ page. The ISOGG health one is particularly worth a read, because it illustrates well some of the analytical challenges with personal health genomics as used today: https://isogg.org/wiki/DNA_in_health_and_disease

Other updates of more technical nature:

Bulk processing is more robust towards data with missing X-chromosomes (this took a lot of time but won’t affect users. Just means fewer pipeline breaks).

Further work with curation of trait types and latest studies (issue #29)

Include filename in the results-ready mail. This will help with multi-genome submissions.

Minor module renaming in precision medicine (it was still called GWAS calculator, a new re-named term, issue #7)

Also issue #9 and #20 are closed now, because they are essentially handled already, at least for people setting up their own servers – one can just dump new prs files in the folder prs_dir and it works.

A bug in Kandinskify yourself caused the pdf download to not follow the on-screen (correct) version, thanks Jarmo!

Update of user-distributions metrics, which is mainly of importance to users that are not of European ancestry.  They’ll now have more accurate background distributions to compare with.

Can now have data folders both as separate volumes and as symlinks. This helps with issue #27